% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{can.template}
\alias{can.template}
\title{Generate templates for canton-level data}
\usage{
can.template(year, endofyear = FALSE, cantons = NULL)
}
\arguments{
\item{year}{a numeric value. Available with GitHub version of package: \code{2001-2017}.}

\item{endofyear}{if \code{FALSE}, cantons as of January 1 are outputted. If \code{TRUE},
cantons as by December 31 are used. End-of-year data is available for \code{year > 2010}.}

\item{cantons}{a numeric (cantons identification numbers) or a character
(two-letter abbreviations) vector to create tailor-made templates.}
}
\description{
\code{can.template} offers the possbility to generate Swiss canton-level data templates in
    the form of data frames. The GitHub version of \code{RSwissMaps} comes with all available geodata
    (\code{2001-2017}). Hence, the \code{map.load} function is not part of this version. The templates
    contain the canton identification numbers as used by the Swiss Federal Statistical Office
    and the canton two-letter abbreviations.
}
\examples{
# Generating template for 2016:
can.template(2016)
}
