% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{coefs}
\alias{coefs}
\alias{ai_dowload_coefs}
\alias{ai_load_coefs}
\alias{ai_unload_coefs}
\title{Coefficients}
\usage{
ai_dowload_coefs(dir = NULL, ...)

ai_load_coefs(dir = NULL)

ai_unload_coefs()
}
\arguments{
\item{dir}{Directory path.}

\item{...}{Arguments parameters passed to [utils::download.file()].}
}
\value{
`ai_dowload_coefs` dowloads a file as a side effect.

`ai_load_coefs` loads the coefficients into the current session.

`ai_unload_coefs` unloads the coefficients from the current session.
}
\description{
Download, load, unload coefficients.
}
\seealso{
[utils::download.file()].
}
