% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dis.plot}
\alias{dis.plot}
\title{Plot thematic district-level maps}
\usage{
dis.plot(bfs_id, data, year, endofyear = FALSE, cantons = NULL,
  districts = NULL, lakes = "all", boundaries = "d",
  boundaries_size = 0.1, boundaries_color = "white", extrema = "global",
  continuous = TRUE, color_continuous = c("#fee5d9", "#a50f15"),
  color_discrete = c("Reds"), color_na = "gray90", title = NULL,
  subtitle = NULL, caption = NULL, legend_title = NULL,
  legend_position = "bottom", save = FALSE, filename = "dis_plot.png",
  dpi = 300, width = NA, heigth = NA, units = "cm")
}
\arguments{
\item{bfs_id}{a numeric vector with district identification numbers as used by the Swiss Federal
Statistical Office corresponding to \code{data}. For templates, see: \code{\link{dis.template}}.}

\item{data}{a vector of data at district-level.}

\item{year}{a numeric value that indicates what map version to plot. Available with GitHub version of
package: \code{2001-2017}.}

\item{endofyear}{if \code{FALSE}, administrative divisions as of January 1 are used. If \code{TRUE},
administrative divisions as by December 31 are plotted. End-of-year data is available for
\code{year > 2009}.}

\item{cantons}{a numeric (canton identification numbers) or a character (two-letter abbreviations)
vector to plot districts of selected cantons.}

\item{districts}{a numeric (district identification numbers) or a character (district names) vector
to plot selected districts.}

\item{lakes}{a numeric (lake identification numbers) or a character (lakes names) vector to
plot selected lakes. If \code{"none"}, no lakes are plotted. If \code{"all"}, all
\href{https://github.com/zumbov2/RSwissMaps/blob/master/lakes}{available lakes} are plotted.}

\item{boundaries}{a character vector naming the boundaries to map:\itemize{
\item \code{"d"}: district boundaries (default)
\item \code{"c"}: cantonal boundaries
\item \code{"n"}: national boundary}}

\item{boundaries_size}{a numeric vector that specifies the sizes of the corresponding elements in \code{boundaries}.}

\item{boundaries_color}{a character vector that specifies the colors of the corresponding elements
in \code{boundaries}.}

\item{extrema}{if \code{"local"}, data of districts that are not selected by \code{cantons} and/or
\code{districts} is not considered in the color scaling process.}

\item{continuous}{set to \code{FALSE}, if \code{data} is non-continuous.}

\item{color_continuous}{a character vector with two elements to specifiy the plot color if \code{continuous =
TRUE}. \code{color_continuous = c("low", "high")} resulting in a two color gradient from color
\code{"low"} to color \code{"high"}.}

\item{color_discrete}{name of a RColorBrewer palette to specifiy the plot color if \code{continuous = FALSE}.
See: \href{https://rpubs.com/flowertear/224344 }{RColorBrewer palettes}.}

\item{color_na}{a character string. color of districts with missing values.}

\item{title, subtitle, caption, legend_title}{text for titles and caption below the plot.}

\item{legend_position}{position of the legend: one of \code{"none"}, \code{"left"}, \code{"right"},
\code{"bottom"}, \code{"top"})}

\item{save}{if \code{TRUE}, the map will be saved to disk.}

\item{filename}{file name to create on disk incl. image file format (".jpeg", ".png", ".tiff", etc.).}

\item{dpi}{plot resolution. Applies only to raster output types.}

\item{width, heigth, units}{plot size in units (\code{"in"}, \code{"cm"}, or \code{"mm"}). If not supplied, uses
the size of current graphics device.}
}
\description{
\code{dis.plot} offers the possbility to visualise Swiss district-level data on a map. The GitHub version
    of \code{RSwissMaps} comes with all available geodata (\code{2001-2017}). Hence, the \code{map.load} function
    is not part of this version. The function arguments make it easy to plot and save customised maps. The function
    \code{\link{mun.template}} provides templates for data collection.
}
\examples{
 # Generating sample data for 2016:
 dt <- dis.template(2016)
 for(i in 1:nrow(dt)){dt$values[i] <- sample(c(300:700), 1)/1000}

 # Plotting sample data of 2016:
 dis.plot(dt$bfs_nr, dt$values, 2016)

 # Plotting sample data for the canton of Aargau:
 dis.plot(dt$bfs_nr, dt$values, cantons = c("AG"), 2016,
 lakes = c("Hallwilersee"))
}
