% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{dis.template}
\alias{dis.template}
\title{Generate templates for district-level data}
\usage{
dis.template(year, endofyear = FALSE, cantons = NULL, districts = NULL)
}
\arguments{
\item{year}{a numeric value. Available with GitHub version of package: \code{2001-2017}.}

\item{endofyear}{if \code{FALSE}, districts as of January 1 are outputted. If \code{TRUE},
districts as by December 31 are used. End-of-year data is available for \code{year > 2009}.}

\item{cantons}{a numeric (canton identification numbers) or a character (two-letter abbreviations)
vector to create canton-specific templates.}

\item{districts}{a numeric (districts identification numbers) or a character
(districts names) vector to create tailor-made templates.}
}
\description{
\code{dis.template} offers the possbility to generate Swiss district-level data templates in
    the form of data frames. The GitHub version of \code{RSwissMaps} comes with all available geodata
    (\code{2001-2017}). Hence, the \code{map.load} function is not part of this version. The templates
    contain the district identification numbers as used by the Swiss Federal Statistical Office
    and the district names.
}
\examples{
# Generating template for 2016:
dis.template(2016)

# Generating template for the municipalities of the canton of Aargau:
dis.template(2016, cantons = c("AG"))
}
