% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_x_text_angle.R
\name{get_x_text_angle}
\alias{get_x_text_angle}
\title{Check the level of column and return the proper angle of X_text_axis}
\usage{
get_x_text_angle(levels)
}
\arguments{
\item{levels}{the level number of a column}
}
\value{
the proper angle of X_text_axis
}
\description{
Check the level of column and return the proper angle of X_text_axis
}
