% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helper}
\alias{helper}
\alias{ai_species}
\alias{ai_classes}
\title{Helpers
Helper functions to list species and vegetation/soil/footprint classes.}
\usage{
ai_species()

ai_classes()
}
\description{
Helpers
Helper functions to list species and vegetation/soil/footprint classes.
}
