% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{mapCH2013}
\alias{mapCH2013}
\title{Geodata of Swiss municipalities, districts, cantons and lakes (2013)}
\format{A data frame with 171,307 rows and 10 variables:
\describe{
  \item{year}{year of geodata}
  \item{endofyear}{if \code{0}, the data point dates from 2013-1-1, else from 2013-31-12}
  \item{long}{longitude of point (Swiss grid CH1903)}
  \item{lat}{latitude of point (Swiss grid CH1903)}
  \item{group}{if two adjacent points are in the same group, then they get connected}
  \item{type}{type of object point belongs to: \code{municipality}, \code{district}, \code{canton},
  \code{country}, \code{lake})}
  \item{id}{official identification number of object}
  \item{name}{name of object}
  \item{dis}{district object belongs to}
  \item{can}{canton object belongs to}}}
\source{
Swiss Federal Statistical Office, GEOSTAT (last download 2017-09-07)
}
\usage{
mapCH2013
}
\description{
A ggplot2-compatible data frame containing geodata of Swiss municipalities, districts,
 cantons and lakes. Reference date: 2013-1-1 and 2013-31-12. The data is made publicly available
 by the \href{https://www.bfs.admin.ch/bfs/de/home/dienstleistungen/geostat/geodaten-bundesstatistik/administrative-grenzen/generalisierte-gemeindegrenzen.html}{Swiss Federal Statistical Office}.
 The data is scaled at ~1:2,000,000.
}
\references{
\url{https://www.bfs.admin.ch/bfs/de/home/dienstleistungen/geostat/geodaten-bundesstatistik/administrative-grenzen/generalisierte-gemeindegrenzen.html}
}
\author{
David Zumbach \email{david.zumbach@gfzb.ch}
}
\keyword{datasets}
