% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{mun.template}
\alias{mun.template}
\title{Generate templates for municipality-level data}
\usage{
mun.template(year, endofyear = FALSE, cantons = NULL, districts = NULL,
  municipalities = NULL)
}
\arguments{
\item{year}{a numeric value. Available with GitHub version of package: \code{2001-2017}.}

\item{endofyear}{if \code{FALSE}, municipalities as of January 1 are outputted. If \code{TRUE},
municipalities as by December 31 are used. End-of-year data is available for \code{year > 2009}.}

\item{cantons}{a numeric (canton identification numbers) or a character (two-letter abbreviations)
vector to create canton-specific templates.}

\item{districts}{a numeric (district identification numbers) or a character (district name)
vector to create district-specific templates.}

\item{municipalities}{a numeric (municipality identification numbers) or a character
(municipality names) vector to create tailor-made templates.}
}
\description{
\code{mun.template} offers the possbility to generate Swiss municipality-level data templates in
    the form of data frames. The GitHub version of \code{RSwissMaps} comes with all available geodata
    (\code{2001-2017}). Hence, the \code{map.load} function is not part of this version. The templates
    contain the municipality identification numbers as used by the Swiss Federal Statistical Office
    and the municipality names.
}
\examples{
# Generating template for 2016:
mun.template(2016)

# Generating template for the municipalities of the canton of Aargau:
mun.template(2016, cantons = c("AG"))
}
