% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{preds}
\alias{preds}
\alias{ai_predict}
\title{Predict}
\usage{
ai_predict(spp, spclim, veghf = NULL, soilhf = NULL, i = 1)
}
\arguments{
\item{spp}{Stecies ID.}

\item{spclim}{Space climate data as a data frame.}

\item{veghf, soilhf}{Vegetation/Soil and footprint info as
(1) a vector of land cover classes or as a
(2) composition matrix with land cover classes as columns.}

\item{i}{Bootstrap ID (1-100).}
}
\description{
Predict based on input data.
}
