% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Put_csv_last.R
\name{put_csv_last}
\alias{put_csv_last}
\title{Put CSV file last in the list}
\usage{
put_csv_last(filelist)
}
\arguments{
\item{filelist}{a list of file names}
}
\value{
a list of file names which the csv files located at the end of it
}
\description{
Put CSV file last in the list
}
