% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helper}
\alias{helper}
\alias{ai_species}
\title{Helpers
Helper functionsDownload, load, unload coefficients.}
\usage{
ai_species()
}
\description{
Helpers
Helper functionsDownload, load, unload coefficients.
}
